Add a krb5_db_check_pkinit_binding() method to kdb
Add a callback for kdb modules which will allow them to return a boolean
indicating whether or not a passed-in client certificate (as a single
binary blob) corresponds to the principal of a particular kdb entry.
We'll call it as a fallback if the PKINIT server plugin doesn't find the
expected principal name in the certificate.