Use krb5_db_check_pkinit_binding() in the KDC
If we can't find a matching SAN in the client's certificate, call
krb5_db_check_pkinit_binding() to check if the supplied certificate
matches the client in question.  If the kdb method doesn't return TRUE,
then we've failed.  This doesn't allow the kdb method to alter any other
requirements we have for client certificates (i.e., we still have to
trust the issuer, and we still check that before we reach this point).