Fix draft9 Diffie-Hellman client requests
If we're creating a draft 9 request, and the user set flag_RSA_PROTOCOL
to "no", use Diffie-Hellman key agreement.

We switch the default from explicitly defaulting to using DH to a having
a per-protocol-version default to keep the traditional behavior if the
user doesn't specifically request one or the other.