Don't try to memset() a NULL buffer
We've gotten a report that when the derived key that's used for
encrypting the TGS reply is detected as being weak (KRB5DES_WEAK_KEY),
krb5_encrypt_helper() fails, krb5_encode_kdc_rep() fails, and
reply.enc_part.ciphertext is left with a nonzero length but with its
data pointer set to NULL.  At that point, the KDC's attempt to clear
these buffers causes it to crash.

Replace the memset()/free() with zapfree(), which should do the same
work, but which also checks the pointer first.