Make the KDC monitor also reopen logs on SIGHUP
In single-process mode, if the KDC receives SIGHUP, it logs a debug
message and reopens its logs.  This is typically done when the logs are
being rotated and a full restart is not desirable.

When the KDC is a monitor for multiple workers, it dutifully passes
SIGHUP on to each of them, which in turn log debug messages and reopen
their logs, but the monitor process hasn't been doing so itself.